Move to QProcesses with custom ProcessManager (#65)
Instead of QThreads running plumbum commands, this PR introduces QProcesses which are used for long running commands (like cloning git repos or building VaRA).
In contrast to QThreads, QProcesses can be terminated reliably.
This is especially important, if the user closes the application during such a long running task.
In this case, the processes are terminated correctly and the application doesn't hang any more.
This is implemented with a custom ProcessManager class with manages the processes.

resolves se-passau/VaRA#271