Show failed projects in vara-cs status (#81)
A *.failed file is created in the results directory if the
analysis of the project fails.
Currently, the "failed" file is only created if the opt command
fails. If the analysis fails at an earlier stage, the failed file
is not created.

The vara-cs status command interprets this file if present, and shows the analysed revision
as failure. If no yaml report and no failure file is present it prints "missing".