Add support for supplementary result files (#129)
This PR enables experiments to create supplementary
result files in addition to the main result file.

The CommitReport.get_supplementary_file_name method
can be used to generate a supplementary report file.
Multiple supplementary files per revision are also
possible.
They can be differentiated by the methods info_type
parameter.

The experiment can then use the generated filenames
to create these files and store additional information
or results of the experiment (e.g. the experiments duration
time).

Contrary to the "normal" result files, these supplementary
files do not indictate a success or failure of the experiment.
They are intended to provide additional information about
successful experiment runs.

This means that if an experiment fails, supplementary files
that have already been created should be deleted.
This can easily be done by the experiment by passing a
list of files to the experiments PEErrorHandler.

If the experiment fails and the error handler is called,
it will automatically delete all files of the passed list.