Adds extender strategy to case study yaml files (#140)
This PR adds the extender strategy and, if available, the sampling method that were used to create/extend each stage of a case study to the case study yaml file format.

If the same stage is extended multiple times with different strategies, the result is mixed as the revisions in the stage cannot be mapped to the different extender strategies any more.
Also, if the same stage is extended by the distribution sampler more than once, the result is always mixed, even if the same distribution was used. This is because the set union of two samples sampled from the same population with the same distribution does in general not result in a valid sample of that distribution.