Taint propagation experiment (#134)
Allow to run vara's taint analysis on designed set of test files for propagation evaluation.

The compiled cpp files and their filecheck files can be found under this repo.

The result of each filecheck is stored into a result file for each test file.
If the output of vara matches the expected filecheck an empty file with the _success file ending is created, otherwise an _failed.txt file will be created. The content of the failed run's output is stored in the result file.