Release extender strategy (#145)
This PR adds a new case study extender that adds revisions identified as a release to the case study. Closes se-passau/VaRA#437

Releases can be further restricted to only major, minor or patch releases with the finer grained release types including the coarser ones.

The extender relies on the project to identify relevant commits. An example implementation for this mechanism is provided for the gzip project.