Allows to mock the config during tests. (#297)
This PR adds a safe way to modify the vara config during tests.
This is achieved by replacing the config with a copy that is tuned for testing using a context manager or decorator. This also allows to write tests that need access to a more-or-less complete tool-suite and benchbuild setup.