New plot interface/CLI prototype.
PlotGenerators are responsible for creating one or more plots.
They get instantiated with a general plot config (as intended by se-sic/VaRA#545) and Plot(Generator) specific options.
Generators also declare these options as CLI options. Common options are provided as constants by the PlotGenerator class.
There is also a helper function to create such CLI option objects.
This separates the implementation detail of the CLI from the plot generation (at least if we further extend/improve this helper function).

The plot driver discovers all plot generators and constructs CLI a command for each generator from its option declarations.
When executed, the created command instantiates the generator class with the general plot config object (it is not yet decided/implemented how this is created) and the generator-specific options.
It then invokes the generator's `__call__()` function passing the remaining arguments that are common to all plots (e.g., the 'view' option).
The idea is that subclasses of PlotGenerator only instantiate the concrete plots with their specific options and, in general, should not need the common options or override the `__call__()` function.

As an example, the paper config overview plot is already compatible with the new architecture.