Decouple artefacts and plots/tables.
This gets rid of that nasty import cycle once and for all by moving concrete artefact implementations to the plots and tables modules and installing our usual __init_subclass__() based register mechanism so that the artefacts module does not need to import anymore. The discovery of artefact implementations can be triggered if needed using the initialize_artefacts() method but usually, those the relevant modules have already been imported somewhere, so initialization will most of the time happen "by accident".