Replace revision strings with CommitHash types (#543)
Starts the replacement of revision strings with CommitHash types by adapting the databases and the interaction_degree_plot as first plot. Further adaptations will be performed in the PlotRework PR (se-sic/VaRA-Tool-Suite#453) or in one of its sub-PRs.

The current state is that most plots are broken, which should be resolved after the PlotRework PR is merged.