Reduce memory overhead of TEFReport (#721)
Since currently the TEFReport loads the whole JSON file into main memory before parsing it, it is easy to run out of memory for huge tracefiles (for example, the main memory usage for a 30 GB tracefile was >200 GB). I implemented a TEFReport that streams the JSON file and, additionally, reduces saving duplicate information by introducing a key-value store that stores function/region names. This way, TraceEvents only need to include an int instead of a (potentially large) string.

Note that I merged #720 into my branch such that the tests do not fail, so not all changes in this branch are related to the above description.