Switches year iteration order in revs_per_year cs extender. (#819)
In Python, random.sample() results seem to depend on the size of the sampled range and the number of drawn samples.
This caused the sample to become unstable when a repo adds new commits. This change should ensure that the sample for all years but the current one remain stable.