Adds lock to FM provider (#882)
When running experiments on slurm, it can happen that multiple nodes fetch the feature model repository at the same time. As nodes may use a shared feature model repository (E.g. on /scratch/) this can lead to issues.

This is no issue if the feature model repository is already up-to-date, however if there are any upstream changes this leads to errors in the fetch on some nodes and subsequent failure of the slurm jobs.


---------

Co-authored-by: Lukas Abelt <git@login.luabelt.de>
Co-authored-by: Florian Sattler <vuld3r@gmail.com>