Improve handling of git operations (#890)
Our handling of git operations in the tool-suite is a mess; some functions use pygit2, others use command-line git and all have different interfaces and conventions and I am very annoyed that I always have the incorrect parameters at hand when I need to use some git functionality.

This PR aims to improve this situation by introducing a lightweight RepositoryHandle class that allows easy and uniform access to both, pygit2 and command-line git`. Since git is ubiquitous in the tool-suite, there are a lot of changes, but most of them are very minor.

As a result of the refactoring, git_util no longer depends on project_util, but the other way round.
In my opinion this makes more sense since projects make use of git, while git can happily be used without projects.

One open point that is remains is to sort out what belongs in git_util vs. git_commands since that is currently not very strict and at least one function in git_commands uses a function in git_util, which imo it should not do.