Adds benchbuild as known third-party in isort.cfg (#922)
This fixes the known_first_party key and adds benchbuild as a known third party. In some local setups (mine) this seems to be required as otherwise isort sorts imports incorrectly on the pre-commit hook. (In my case, specifically and only import benchbuild as bb always gets put in a wrong position)

Co-authored-by: Lukas Abelt <git@login.luabelt.de>